Getting rid of reflection Utils
