--- llap-server/src/test/org/apache/hadoop/hive/llap/TestReflectionUtils.java
@@ -1,54 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.hadoop.hive.llap;
-
-import static org.junit.Assert.fail;
-
-import java.lang.reflect.Field;
-
-public class TestReflectionUtils {
-
-  public static Field reflectField(Class<?> classToReflect, String fieldNameValueToFetch) {
-    try {
-      Field reflectField = null;
-      Class<?> classForReflect = classToReflect;
-      do {
-        try {
-          reflectField = classForReflect.getDeclaredField(fieldNameValueToFetch);
-        } catch (NoSuchFieldException e) {
-          classForReflect = classForReflect.getSuperclass();
-        }
-      } while (reflectField==null || classForReflect==null);
-      reflectField.setAccessible(true);
-      return reflectField;
-    } catch (Exception e) {
-      fail("Failed to reflect " + fieldNameValueToFetch + " from " + classToReflect);
-    }
-    return null;
-  }
-
-  public static void reflectSetValue(Object objToReflect, String fieldNameToSet, Object valueToSet) {
-    try {
-      Field reflectField  = reflectField(objToReflect.getClass(), fieldNameToSet);
-      reflectField.set(objToReflect, valueToSet);
-    } catch (Exception e) {
-      fail("Failed to reflectively set " + fieldNameToSet + "=" + valueToSet);
-    }
-  }
-
-}
